api/src/requirements.txt
python-dotenv==1.2.1
fastapi==0.128.0
uvicorn==0.40.0
psycopg2-binary==2.9.11
email-validator>=2.0
sqlalchemy==2.0.46
alembic==1.18.3